Add files via upload
Using STM32 microcontroller with button for activating the diod blinking. the model was made in the Proteus professional environment. main.C  file and diode.hex was writen in the STM32CubeIDE. Download the STM32_Button.pdsprj file, open the STM32 model and and specify the path to the .hex file. Then, start simulation by the left down button. LED should blinking with the 500ms. frequency after pressing the button.